+++ src/main/kotlin/insyncwithfoo/ryecharm/ruff/documentation/RuleDocumentation.kt
@@ -2,6 +2,7 @@ package insyncwithfoo.ryecharm.ruff.documentation
 
 import com.intellij.openapi.application.readAction
 import com.intellij.openapi.project.Project
+import insyncwithfoo.ryecharm.DocumentationURI
 import insyncwithfoo.ryecharm.HTML
 import insyncwithfoo.ryecharm.Markdown
 import insyncwithfoo.ryecharm.ProgressContext
@@ -10,6 +11,31 @@ import insyncwithfoo.ryecharm.processTimeout
 import insyncwithfoo.ryecharm.ruff.commands.ruff
 import insyncwithfoo.ryecharm.runInBackground
 import insyncwithfoo.ryecharm.toHTML
+
+
+private val optionsSection = """(?mx)
+    ^\#\#\h*Options\n
+    (?:[\s\S](?!\n\#))*
+""".trimIndent().toRegex()
+
+
+private val optionNameInListItem = """(?m)(?<prefix>^[-*]\h*\[?)`(?<path>[A-Za-z0-9.-]+)`""".toRegex()
+
+
+// https://github.com/astral-sh/ruff/issues/14348
+/**
+ * Replace option names with links to specialized URIs.
+ */
+private fun String.insertOptionLinks() = this.replace(optionsSection) {
+    it.value.replace(optionNameInListItem) { match ->
+        val prefix = match.groups["prefix"]!!.value
+        val path = match.groups["path"]!!.value
+        
+        val uri = DocumentationURI(RUFF_OPTION_HOST, path)
+        
+        "$prefix[`$path`]($uri)"
+    }
+}
 
 
 private suspend fun Project.getMarkdownDocumentationForRule(rule: String): Markdown? {
@@ -36,5 +62,5 @@ private suspend fun Project.getMarkdownDocumentationForRule(rule: String): Markd
 internal suspend fun Project.getDocumentationForRule(rule: String): HTML? {
     val markdownDocumentation = getMarkdownDocumentationForRule(rule) ?: return null
     
-    return readAction { markdownDocumentation.toHTML() }
+    return readAction { markdownDocumentation.insertOptionLinks().toHTML() }
 }